chore: merge user management branch changes and fix imports and variable names

+++ src/main/java/com/group4/gamehub/dto/PublicUserResponse.java
@@ -1,6 +1,7 @@
 package com.group4.gamehub.dto;
 
-import com.group4.gamehub.model.Role;
+import com.group4.gamehub.util.Role;
+
 import lombok.Builder;
 import lombok.Getter;
 
@@ -10,5 +11,5 @@ public class PublicUserResponse {
     private String username;
     private Role role;
     private String rank;
-    private int points;
+    private Long points;
 }
