+++ src/main/java/com/group4/gamehub/dto/UserResponse.java
@@ -1,6 +1,6 @@
 package com.group4.gamehub.dto;
 
-import com.group4.gamehub.model.Role;
+import com.group4.gamehub.util.Role;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -12,5 +12,5 @@ public class UserResponse {
     private String email;
     private Role role;
     private String rank;
-    private int points;
+    private Long points;
 }
